backend/souvenir-microservice/requirements.txt
connexion[swagger-ui]==2.6.0
openapi_spec_validator==0.4.0
gunicorn==20.0.4
requests==2.22.0
flask-cors==3.0.10
Pillow==8.4.0
opencv-python
torch==1.12.1
torch-vision
Cython
insightface
onnxruntime
psutil
uvicorn
google-cloud
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
waitress

